% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_data.R
\docType{data}
\name{default_cortical_measures}
\alias{default_cortical_measures}
\title{Default cortical measures}
\format{
An object of class \code{character} of length 9.
}
\usage{
default_cortical_measures
}
\description{
The subset of cortical bone parameters reported by default:
\code{Ct.Ar}, \code{Ma.Ar}, \code{Tt.Ar}, \code{Ct.Ar/Tt.Ar}, \code{Ct.Th}, \code{Ct.Po}, \code{Ps.Pm},
\code{Es.Pm}, and \code{Ct.vBMD}.
}
\keyword{datasets}
