% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_data.R
\docType{data}
\name{default_trabecular_measures}
\alias{default_trabecular_measures}
\title{Default trabecular measures}
\format{
An object of class \code{character} of length 4.
}
\usage{
default_trabecular_measures
}
\description{
The subset of trabecular bone parameters reported by default:
\code{Tb.BV/TV}, \code{Tb.Th}, \code{Tb.Sp}, and \code{Tb.N}.
}
\keyword{datasets}
