% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{klct_ex}
\alias{klct_ex}
\title{Get path to klct examples}
\usage{
klct_ex(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\value{
A string containing the name of the file, or a character vector
containing the names of all example files.
}
\description{
klct comes bundled with some example files in its \code{inst/extdata}
directory. This function makes them easy to access.
}
\details{
Adapted from \code{\link[readxl:readxl_example]{readxl::readxl_example()}}.
}
\examples{
klct_ex()
klct_ex("mouse-table.csv")
}
