% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_data.R
\docType{data}
\name{metadata_cols}
\alias{metadata_cols}
\title{Non-measure metadata columns excluded during detection}
\format{
An object of class \code{character} of length 8.
}
\usage{
metadata_cols
}
\description{
Non-measure metadata columns excluded during detection
}
\keyword{internal}
