% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_results.R
\name{print_data}
\alias{print_data}
\title{Print bone microCT data}
\usage{
print_data(data, ...)
}
\arguments{
\item{data}{A data frame containing bone microCT data, formatted as is the
output of \code{\link[=read_trabecular]{read_trabecular()}} or \code{\link[=read_cortical]{read_cortical()}}.}

\item{...}{Additional arguments passed on to \code{\link[DT:datatable]{DT::datatable()}}.}
}
\value{
An \code{\link[htmltools:tagList]{htmltools::tagList()}} containing one \code{\link[DT:datatable]{DT::datatable()}} widget
per sex.
}
\description{
This is a wrapper around \code{\link[DT:datatable]{DT::datatable()}} to print out bone microCT
data at one site by sex for use in an R Markdown document. Each sex
is displayed as an interactive, sortable, searchable table.
}
\examples{
mouse_table <- read_mouse_table(klct_ex("mouse-table.csv"))
spine_trab <- read_trabecular(klct_ex("spine_trabecular_data.csv"),
                              mouse_table, site = "Spine")
print_data(spine_trab)
}
