% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{print_plots}
\alias{print_plots}
\title{Print a list of plots}
\usage{
print_plots(plots)
}
\arguments{
\item{plots}{A list of \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} objects.}
}
\value{
Returns NULL, since this function is used for its side effects.
}
\description{
This function iterates through a list of plots and prints them (without
printing the list index).
}
\examples{
mouse_table <- read_mouse_table(klct_ex("mouse-table.csv"))
met_cort <- read_cortical(klct_ex("femur_metaphysis_data.csv"),
                          mouse_table, site = "Met")
plot_groups(met_cort) |> print_plots()
}
