% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_results.R
\name{print_results}
\alias{print_results}
\title{Print bone microCT comparison analysis results}
\usage{
print_results(results, sig_color = "red", ...)
}
\arguments{
\item{results}{A list of microCT comparison results, formatted as is the
output of \code{\link[=compare_groups]{compare_groups()}}.}

\item{sig_color}{What color to print a significant result. Defaults to
\code{"red"}. Set to \code{NULL} to disable (and just print everything black).}

\item{...}{Additional arguments passed on to \code{\link[knitr:kable]{knitr::kable()}}.}
}
\value{
Text output which is by default a Pandoc markdown pipe table for each
measure and each sex analyzed. If two sexes are analyzed, Pandoc fenced div
syntax is used to print the tables in two columns by sex.
}
\description{
This function wraps and extends \code{\link[knitr:kable]{knitr::kable()}} to format and print the
results of bone microCT comparison analyses at one site by sex for use in an
R Markdown document. A code chunk containing this function should probably
have the \code{results = "asis"} option to allow pretty formatting. Any \code{NA}
values are printed as an empty string.
}
\examples{
mouse_table <- read_mouse_table(klct_ex("mouse-table.csv"))
spine_trab <- read_trabecular(klct_ex("spine_trabecular_data.csv"),
                              mouse_table, site = "Spine")
spine_res <- compare_groups(spine_trab)
print_results(spine_res)
}
