% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_mouse_table}
\alias{read_mouse_table}
\title{Read in a mouse table}
\usage{
read_mouse_table(file, ...)
}
\arguments{
\item{file}{The file path to the mouse table CSV file.}

\item{...}{Additional arguments passed on to \code{\link[readr:read_delim]{readr::read_csv()}}.}
}
\value{
A tibble/data frame containing the mouse definitions. For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{| Animal ID|   AS|Sex    |Treatment |
|:---------|----:|:------|:---------|
| JQ10.L1  | 1579|Female |AP        |
| JQ10.2   | 1580|Female |AP        |
| JQ11.2   | 1581|Male   |AP        |
}\if{html}{\out{</div>}}
}
\description{
This is a wrapper around \code{\link[readr:read_delim]{readr::read_csv()}} to read in and process a mouse
table CSV file that contains the lookup table defining each sample's Animal
ID, AS (key registry) number, sex, and treatment.
}
\examples{
mouse_table <- read_mouse_table(klct_ex("mouse-table.csv"))
}
