% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_trabecular}
\alias{read_trabecular}
\title{Read in trabecular bone data}
\usage{
read_trabecular(file, mouse_table, site, ...)
}
\arguments{
\item{file}{The file path to the trabecular data CSV file. This can be either
the spine trabecular data file or the femur metaphysis data file (which
contains both cortical and trabecular columns).}

\item{mouse_table}{The mouse table object containing sample information, as
created by the \code{\link[=read_mouse_table]{read_mouse_table()}} function.}

\item{site}{A string indicating the anatomical site. Options include
\code{"Spine"}, \code{"Met"} (femoral metaphysis), or any custom site label. This is
added as a \code{Site} column to the output.}

\item{...}{Additional arguments passed on to \code{\link[readr:read_delim]{readr::read_csv()}}.}
}
\value{
A tibble/data frame containing trabecular bone data. For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{|   AS|Sex |Treatment |Site  | Tb.BV/TV| Tb.Th| Tb.Sp| Tb.N|
|----:|:---|:---------|:-----|--------:|-----:|-----:|----:|
| 1579|F   |AP        |Spine |    18.69| 0.032| 0.252| 3.52|
| 1580|F   |AP        |Spine |    21.60| 0.045| 0.292| 2.97|
}\if{html}{\out{</div>}}
}
\description{
This function reads in trabecular bone microCT data from a NeoScan CSV file
and joins it with a mouse table to add sex and treatment information.
}
\details{
For femoral metaphysis data, the trabecular columns are extracted from the
combined cortical/trabecular file. For spine data, the file contains only
trabecular columns.
}
\examples{
mouse_table <- read_mouse_table(klct_ex("mouse-table.csv"))
spine_trab <- read_trabecular(klct_ex("spine_trabecular_data.csv"),
                              mouse_table,
                              site = "Spine")
}
