% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_klct}
\alias{update_klct}
\title{Update klct}
\usage{
update_klct()
}
\description{
This is a convenience function wrapper around \code{\link[remotes:install_github]{remotes::install_github()}} to
install/update \code{klct}. Meant to be used as an RStudio addin.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
update_klct()
\dontshow{\}) # examplesIf}
}
