% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_results.R
\name{print_data}
\alias{print_data}
\title{Print bone microCT data}
\usage{
print_data(data, ...)
}
\arguments{
\item{data}{A data frame containing bone microCT data, formatted as is the
output of \code{\link[=read_trabecular]{read_trabecular()}} or \code{\link[=read_cortical]{read_cortical()}}.}

\item{...}{Additional arguments passed on to \code{\link[knitr:kable]{knitr::kable()}}.}
}
\value{
Text output which by default is a Pandoc markdown pipe table for each
sex contained in the \code{data} supplied.
}
\description{
This is a wrapper around \code{\link[knitr:kable]{knitr::kable()}} to print out bone microCT
data at one site by sex for use in an R Markdown document. A code
chunk containing this function should probably have the \code{results = "asis"}
option set to allow for pretty formatting.
}
\examples{
mouse_table <- read_mouse_table(mctr_ex("CD2MouseTable.csv"))
spine_trab <- read_trabecular(mctr_ex("spine_trabecular_data.csv"),
                              mouse_table, site = "Spine")
print_data(spine_trab)
}
